requirements.txt
flask
requests
python-dotenv
pytest
pytest-mock
bleach
gunicorn
